feat: rework Bind & Binder
Bind is no longer a collection of ToSql types, instead every ToSql type
is Bind, and a collection of Bind types is also Bind.
This recursive definition fits better with moving Tables from being a
collection of Columns to being a collection of Values.

Additionally, the Binder holds the Statement by (mut) reference instead
of by value. This way it's easier to use, it can just be dropped when we
need the Statement back.