feat: add CloneFromRef trait to extend Tuple
Unfortunately it does not seem possible to integrate this into the Tuple
trait itself: although it is possible to add methods not supported by
all implementors (e.g. only where Self: Clone) the trait system pretends
not to know that for any (A, B), Self: Clone IFF A: Clone and B: Clone.
Because of this it is not possible to specify a bound in the trait
declaration such that all cases are covered.